Refactor main.cpp by extracting app layer modules
- Extract crash handling to app/crash_handler with Linux-specific implementation
- Extract logger setup and argument parsing to app/app_setup
- Reduce main.cpp from 205 lines to 64 lines for better maintainability
- Add cross-platform support with graceful fallbacks for non-Linux systems
